+++ camera.go
@@ -17,13 +17,17 @@ func NewCamera(windowWidth, windowHeight int) (*Camera, error) {
 	return &Camera{windowWidth / 2, windowHeight / 2, windowWidth, windowHeight, 10}, nil
 }
 
-func (c *Camera) SetToPosition(x int, y int) {
-	c.x = x + c.windowWidth/2
-	c.y = y + c.windowHeight/2
+func (c *Camera) SetToPosition(x, y int) {
+	c.x = c.windowWidth / 2 + x
+	c.y = c.windowHeight / 2 + y
 }
 
 func (c *Camera) IsVisible(x, y int) bool {
-	return !(x < 0 || x > c.windowWidth || y < 0 || y > c.windowHeight)
+	lowWidth := x < c.windowWidth / 2 - c.x
+	hightWidth := x > c.windowWidth / 2 + c.x
+	lowHeight := y < c.windowHeight / 2 - c.y
+	highWidth := y > c.windowHeight / 2 + c.y
+	return !(lowWidth || hightWidth || lowHeight || highWidth)
 }
 
 func (c *Camera) Zoom(amount int16) {