+++ drawing.go
@@ -8,7 +8,7 @@ const TWO_PI = math.Pi * 2
 var defaultColor = Color{Red: 127, Green: 127, Blue: 127, Alpha: 255}
 
 func DrawCircle(renderer Renderer, x, y, radius int, color Color) {
-	// Object is too small
+	// Object is too small, draw in default color
 	if radius == 0 {
 		renderer.SetDrawColor(defaultColor.Red, defaultColor.Green, defaultColor.Blue, defaultColor.Alpha)
 		renderer.DrawPoint(int(x), int(y))